@@ -1,34 +1,33 @@
 {
-  "name": "@ese_tecnodigital/monitor-covid",
+  "name": "huddle",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@ese_tecnodigital/dashboard": "^0.1.15",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "react": "^16.13.1",
-    "react-csv": "^2.0.3",
-    "react-dom": "^16.13.1",
-    "react-icons": "^3.10.0",
-    "react-redux": "^7.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.1",
-    "react-spring": "^8.0.27",
-    "react-timeago": "^4.4.0",
-    "recharts": "^1.8.5",
-    "redux": "^4.0.5",
-    "socket.io-client": "^4.2.0",
-    "toasted-notes": "^3.2.0"
+    "@emotion/react": "^11.9.3",
+    "@emotion/styled": "^11.9.3",
+    "@mui/material": "^5.9.1",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^13.5.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.4.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
+    "react-table": "^7.8.0",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-app-rewired start --scripts-version react-scripts",
-    "build": "react-app-rewired build --scripts-version react-scripts",
-    "test": "react-app-rewired test --scripts-version react-scripts",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
@@ -41,8 +40,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "react-app-rewired": "^2.1.6"
   }
 }
