@@ -1,19 +1,22 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import { Provider } from 'react-redux';
-
+import ReactDOM from 'react-dom/client';
+import './index.css';
 import App from './App';
-import store from './store';
+import reportWebVitals from './reportWebVitals';
 
-ReactDOM.render(
-  <Provider store={store}>
-    <React.StrictMode>
-      <App />
-    </React.StrictMode>
-  </Provider>,
-  document.getElementById('root')
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+  <React.StrictMode>    
+    <App />
+  </React.StrictMode>
 );
 
-if (module.hot) {
-  module.hot.accept();
-}
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals();
+
+
+root.render(
+ 
+);
