@@ -1,43 +1,23 @@
-import React from 'react';
-import { connect } from 'react-redux';
-import { BrowserRouter, Route, Switch } from 'react-router-dom';
+import React from "react";
+import { Route, BrowserRouter, Routes } from "react-router-dom";
+import Home from "./pages/Home";
+import Admin from "./pages/Admin";
+import Login from "./pages/Login";
+import Header from "./components/Header"
 
-import { AiOutlineHome, AiOutlineBarChart, AiOutlineHistory } from 'react-icons/ai';
-import { Header } from '@ese_tecnodigital/dashboard';
-import Monitor from './pages/Monitor';
-import HospitalBed from './pages/HospitalBed';
-import Charts from './pages/Charts';
-import Reports from './pages/Reports';
-import { sensorCheckData } from './actions';
-import coppeImg from './assets/coppe.png';
-import '@ese_tecnodigital/dashboard/dist/index.cjs.css'
-
-const Routes = ({sensorCheckData}) => {
-  return (
+const RoutesApp = () => {
+   return(
     <BrowserRouter>
-      <Header logo={coppeImg} sensorCheckData={sensorCheckData} dashboardName="Oximetro IoT UFRJ" headerItens={[{
-        path: "/",
-        title:"Página Inicial",
-        icon: <AiOutlineHome size={32} />
-      }, {
-        path: "/charts",
-        title:"Gráficos Barra",
-        icon: <AiOutlineBarChart size={32} />
-      }, {
-        path: "/reports",
-        title:"Relatório de Sinais Vitais",
-        icon: <AiOutlineHistory size={32} />
-      }]}
- />
-
-      <Switch>
-        <Route path="/" exact component={Monitor} />
-        <Route path="/beds/:id" component={HospitalBed} />
-        <Route path="/charts" component={Charts} />
-        <Route path="/reports" component={Reports} />
-      </Switch>
+       <Header/>
+       <Routes> 
+            <Route element = {< Home/> }  path="/" exact />        
+            <Route element = { <Admin/> }  path="/admin"  />        
+            <Route element = { <Login/> }  path="/login"  />                   
+        </Routes>
     </BrowserRouter>
-  );
-};
+   )
+}
 
-export default connect(null, { sensorCheckData })(Routes);
+export default RoutesApp;
+
+
